requirements.txt
# RunPod
runpod>=1.6.0

# Core ML libraries
torch>=2.1.0
diffusers>=0.30.0
transformers>=4.40.0
accelerate>=0.27.0
safetensors>=0.4.0

# Image processing
Pillow>=10.0.0

# CUDA optimizations (optional, install if using GPU)
# xformers>=0.0.23  # For memory-efficient attention

# Utilities
pydantic>=2.0.0
